pyproject.toml
[tool.poetry]
name = "http-info-bot"
version = "0.1.0"
description = "Chatbot para consulta de informações de códigos HTTP"
authors = ["Douglas Gusson <douglasgusson@gmail.com>"]
license = "MIT"
readme = "README.md"

repository = "https://github.com/douglasgusson/http-info-bot"
keywords = ["python", "scraping", "bot", "telegram"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.4.1"
dynaconf = "^3.1.4"
beautifulsoup4 = "^4.9.3"
requests = "^2.25.1"
bs4 = "^0.0.1"
emoji = "^1.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
isort = "^5.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
